bus/hp_dio/98550: convert to anonymous namespace (nw)
