+++ src/devices/bus/hp_dio/hp98550.cpp
@@ -18,14 +18,66 @@ HP98550 high-resolution color board
 //#define VERBOSE 1
 #include "logmacro.h"
 
+namespace {
+
+class dio32_98550_device :
+	public device_t,
+	public bus::hp_dio::device_dio32_card_interface,
+	public device_memory_interface
+{
+public:
+	dio32_98550_device(const machine_config &mconfig, const char *tag, device_t *owner, uint32_t clock);
+
+	uint16_t rom_r(offs_t offset, uint16_t mem_mask = ~0);
+	void rom_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+
+	uint16_t catseye_r(address_space &space, offs_t offset, uint16_t mem_mask = ~0);
+	void catseye_w(address_space &space, offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+
+	uint16_t vram_r(offs_t offset, uint16_t mem_mask = ~0);
+	void vram_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
+
+	static constexpr int CATSEYE_COUNT = 8;
+
+
+	uint32_t screen_update(screen_device &screen, bitmap_rgb32 &bitmap, const rectangle &cliprect);
+protected:
+	required_device<nereid_device> m_nereid;
+	required_device_array<catseye_device, CATSEYE_COUNT> m_catseye;
+
+	dio32_98550_device(const machine_config &mconfig, device_type type, const char *tag, device_t *owner, uint32_t clock);
+
+	virtual void device_start() override;
+	virtual void device_reset() override;
+
+	virtual void device_add_mconfig(machine_config &config) override;
+	virtual const tiny_rom_entry *device_rom_region() const override;
+
+	virtual space_config_vector memory_space_config() const override;
+
+	void vblank_w(int state);
+	void int_w(offs_t offset, uint8_t data);
+
+	const address_space_config m_space_config;
+	void map(address_map &map);
+	void update_int();
+
+	static constexpr int m_fb_width = 2048;
+	static constexpr int m_h_pix = 1280;
+	static constexpr int m_v_pix = 1024;
+
+	required_region_ptr<uint8_t> m_rom;
+	required_shared_ptr_array<uint8_t, 2> m_vram;
+
+	uint16_t m_plane_mask;
+	uint8_t m_intreg;
+	uint8_t m_ints;
+};
+
 ROM_START(hp98550)
 	ROM_REGION(0x8000, "hp98550a_rom", 0)
 	ROM_LOAD("98550a.bin", 0x000000, 0x008000, CRC(9d639233) SHA1(d6b23a34850f24525ca5fb36de3deb91196d2dc5))
 ROM_END
-
-DEFINE_DEVICE_TYPE(HPDIO_98550, bus::hp_dio::dio32_98550_device, "dio98550", "HP98550A high-res color DIO video card")
-
-namespace bus::hp_dio {
 
 void dio32_98550_device::device_add_mconfig(machine_config &config)
 {
@@ -70,7 +122,7 @@ dio32_98550_device::dio32_98550_device(const machine_config &mconfig, const char
 
 dio32_98550_device::dio32_98550_device(const machine_config &mconfig, device_type type, const char *tag, device_t *owner, uint32_t clock) :
 	device_t(mconfig, type, tag, owner, clock),
-	device_dio16_card_interface(mconfig, *this),
+	device_dio32_card_interface(mconfig, *this),
 	device_memory_interface(mconfig, *this),
 	m_nereid(*this, "nereid"),
 	m_catseye(*this, "catseye%d", 0),
@@ -232,4 +284,6 @@ uint32_t dio32_98550_device::screen_update(screen_device &screen, bitmap_rgb32 &
 	return 0;
 }
 
-} // namespace bus::hp_dio
+} // anonymous namespace
+
+DEFINE_DEVICE_TYPE_PRIVATE(HPDIO_98550, bus::hp_dio::device_dio32_card_interface, dio32_98550_device, "dio98550", "HP98550A high-res color DIO video card")
